requirements.txt
appnope==0.1.3
asgiref==3.8.1
asttokens==2.4.0
attrs==23.2.0
backcall==0.2.0
cached-property==1.5.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
cryptography==41.0.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.3
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
dnspython==2.4.2
drf-yasg==1.21.7
exchangelib==5.1.0
executing==2.0.0
filelock==3.13.1
fsspec==2023.10.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipython-sql==0.5.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.4.0
jupyter_core==5.4.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.32
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.6
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.25.0
pycparser==2.21
Pygments==2.16.1
PyGObject==3.46.0
pyspnego==0.10.2
pytest==7.4.4
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pywatchman==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.34.0
regex==2023.10.3
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
scikit-surprise==1.1.3
scipy==1.11.4
six==1.16.0
sklearn==0.0.post10
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.3
surprise==0.1
swagger-spec-validator==3.0.3
sympy==1.12
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.8
zmq==0.0.0
